Done SideBar

+++ HaT7FptBook/Areas/StoreOwner/Controllers/BooksController.cs
@@ -267,7 +267,7 @@ namespace HaT7FptBook.Areas.StoreOwner.Controllers
                         Title = reader.GetValue(2).ToString(),
                         Description = reader.GetValue(3).ToString(),
                         Author = reader.GetValue(4).ToString(),
-                        NoPage = reader.GetValue(5).GetHashCode(),
+                        NoPage = (int)reader.GetValue(5).GetHashCode(),
                         Price = (double)reader.GetValue(6),
                         CategoryId = _db.Categories.FirstOrDefault(c => c.Name == reader.GetValue(7)).Id
                     };
